api/requirements.txt
aniso8601==7.0.0
appnope==0.1.2
asgiref==3.4.1
backcall==0.2.0
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
cryptography==3.4.7
decorator==5.0.9
defusedxml==0.7.1
Django==3.2.5
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
django-recaptcha==2.0.6
django-unixtimestampfield==0.4.0
graphene==2.1.8
graphene-django==2.15.0
graphene-django-pagination @ https://github.com/instruct-br/graphene-django-pagination/archive/master.zip
graphql-core==2.3.2
graphql-relay==2.0.1
idna==2.10
ipython==7.16.1
ipython-genutils==0.2.0
jedi==0.17.2
oauthlib==3.1.1
parso==0.7.1
pexpect==4.8.0
pickleshare==0.7.5
promise==2.3
prompt-toolkit==3.0.19
psycopg2==2.9.1
ptyprocess==0.7.0
pycparser==2.20
Pygments==2.9.0
PyJWT==2.1.0
python3-openid==3.2.0
pytz==2021.1
requests==2.25.1
requests-oauthlib==1.3.0
Rx==1.6.1
sentry-sdk==1.3.0
singledispatch==3.6.2
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.1.0
sqlparse==0.4.1
text-unidecode==1.3
traitlets==4.3.3
typing-extensions==3.10.0.0
urllib3==1.26.6
wcwidth==0.2.5
